+++ web/src/components/speaker/CardHeader.vue
