+++ web/src/components/speaker/SpeakerCard.vue
